Finally resolved movement problem
Although some bugs still need to be fixed, player movement is almost finished. From can jump, and stand on the platforms. Falling is made quite smooth. Still need to fix: 1. Player is making a hole when is going through the platform. 2. Sometimes Player stays on the platfrom in the middle of a jump.